Description of game3i.txt
I used the BJ282 software application to do the following. Specify the above-described set of game rules that players agree to at some high stakes blackjack tables found on the Las Vegas strip. Conduct a complete-combinatorial analysis of the card game. Use the results of the complete-combinatorial analysis to assemble a set of combination-dependent-playing-strategy tables for the card game. Use the results of the complete-combinatorial analysis to assemble a combination-dependent-expected-value table for the card game. Use the results of the complete-combinatorial analysis to calculate an estimate of the expected value of the card game. Save the results of the complete-combinatorial analysis to a set of files consisting of a set of data files and a set of text files. 

The set of text files includes a text file named game3i.txt.. This text file contains a record of how I responded to prompts for user input made by the BJ282 software application.