Description of game3ev.txt
I used the BJ282 software application to do the following. Specify the above-described set of game rules that players agree to at some high stakes blackjack tables found on the Las Vegas strip. Conduct a complete-combinatorial analysis of the card game. Use the results of the complete-combinatorial analysis to assemble a set of combination-dependent-playing-strategy tables for the card game. Use the results of the complete-combinatorial analysis to assemble a combination-dependent-expected-value table for the card game. Use the results of the complete-combinatorial analysis to calculate an estimate of the expected value of the card game. Save the results of the complete-combinatorial analysis to a set of files consisting of a set of data files and a set of text files. 

The set of text files includes a text file named game3ev.txt. This text file contains a set of tables. The set of tables consists of a table for each combination of point values that could be assigned to a set of cards consisting of the two cards in the player's initial hand and the card dealt face up in the dealer's initial hand. In each of the set of tables, the BJ282 software application does the following.  Gives the point values assigned the two cards in the player's initial hand.  Gives the point value assigned to the card dealt face up in the dealer's initial hand. Gives the dealer finish probabilities. Gives the percentage of all hands that were accounted for by the complete-combinatorial analysis. Identifies the best strategy for the player's use of the hit option. Gives the expected value of the best strategy for the player's use of the hit option. Gives the expected values of all other options for the play of the player's hand that are permitted by the predetermined set of game rules. Identifies the best strategy for the play of the player's hand.