Description of game3o.txt
I used the BJ282 software application to do the following. Specify the above-described set of game rules that players agree to at some high stakes blackjack tables found on the Las Vegas strip. Conduct a complete-combinatorial analysis of the card game. Use the results of the complete-combinatorial analysis to assemble a set of combination-dependent-playing-strategy tables for the card game. Use the results of the complete-combinatorial analysis to assemble a combination-dependent-expected-value table for the card game. Use the results of the complete-combinatorial analysis to calculate an estimate of the expected value of the card game. Save the results of the complete-combinatorial analysis to a set of files consisting of a set of data files and a set of text files. 

The set of text files includes a text file named game3o.txt. This text file contains a set of tables. The set of tables consists of one combination-dependent-playing-strategy table assembled for the set of hands consisting of the player's primary-initial hand and the dealer's initial hand, a composition-dependent-expected-value table assembled for the set of hands consisting of the player's primary-initial hand and the dealer's initial hand, and two combination-dependent-playing-strategy tables assembled for the set of hands consisting of the player's post-split-initial hand and the dealer's initial hand.