Description of 17h17s6dgame3Ai.txt
I used the ULTIMA283 software application to load data into program memory from a set of data files. The BJ282 software application saved the set of data files. The set of data files includes a data file named game3a.dat. This data file contains data about a predetermined set of game rules for a card game belonging to the Blackjack family. The set of data files includes a data file named game3b.dat. This data file contains data about a set of tables. The BJ282 software application assembled the set of tables for the card game. The set of tables  consists of a set of combination-dependent-playing-strategy tables, and a combination-dependent-expected-value table. 

The ULTIMA283 software application used the data contained in the set of data files to do the following. Specify a predetermined set of game rules for a card game belonging to the Blackjack family. Program a computer-controlled player to use the combination-dependent-playing-strategy tables to make decisions on how to play the player's hand. 

The predetermined set of game rules was a set of game rules that players agree to at some high stakes blackjack tables found on the Las Vegas strip.  The predetermined set of game rules specified the following. The dealer uses a six deck shoe. The dealer stands on soft 17. If the card dealt face up to the dealer's initial hand is not a card selected from a group consisting of an Ace and a card assigned a value of ten points, then the player may surrender the player's initial hand. If the card dealt face up to the dealer's initial hand is a card selected from a group consisting of an Ace and a card assigned a value of ten points,  and if, upon checking the dealer's initial hand for a blackjack,  the dealer does not find a blackjack, then the player may surrender the player's initial hand. The player may split a pair, and re-split a maximum of three times. The player may double down on any two cards. The player may double down after a split. The player may re-split aces. The player may not draw to split aces. Blackjack pays 3 to 2 odds on the game wager. 

I further programmed the ULTIMA283 software application to do the following. Simulate one-billion rounds of the card game. Divide the simulation of one-billion rounds of the card game up into one hundred sessions with each session consisting of the simulation of ten-million rounds of the card game. Have the computer-controlled player bet the table-minimum during each round of the card game. Use on-going-game-simulation results to converge upon an optimal-decimal-card-counting strategy (a.k.a an “ezcount” strategy). Track various other statistics including the expected value of the game. Save the results of the game simulation to a set of files consisting of a set of data files and a set of text files.

The set of text files includes a text file named 17h17s6dgame3Ai.txt. This text file contains a record of how I responded to prompts for user input made by the ULTIMA283 software application. The record can be found here.