Description of 17h17s6dgame3Ao.txt
The set of text files includes a text file named 17h17s6dgame3Ao.txt. The ULTIMA283 software application enables the user to divide up a simulation into a number of sessions. At the end of each session, the ULTIMA283 software application does the following. Updates the application window with data about last round of simulated play in that session. Updates the application window with the various statistics that were tracked during that session. Appends this text file with a copy of the update to the application window. 

In this example, I programmed the ULTIMA283 software application to divide a one-billion round simulation into one hundred sessions of ten million rounds each.