Description of 17h17s6dgame3Ad.txt
I used the ULTIMA283 software application to load data into program memory from a set of data files. The BJ282 software application saved the set of data files. The set of data files includes a data file named game3a.dat. This data file contains data about the above-described set of game rules that players agree to at some high stakes blackjack tables found on the Las Vegas strip. The set of data files includes a data file named game3b.dat. This data file contains data about a set of tables. The BJ282 software application assembled the set of tables for the card game. The set of tables  consists of a set of combination-dependent-playing-strategy tables, and a combination-dependent-expected-value table. 

The ULTIMA283 software application used the data contained in the set of data files to do the following. Specify a predetermined set of game rules for a card game belonging to the Blackjack family of table-card games. Program a computer-controlled player to make decisions on how to play the player's hand. 

I further programmed the ULTIMA283 software application to do the following. Simulate one-billion rounds of the card game. Divide the simulation of one-billion rounds of the card game up into one hundred sessions with each session consisting of the simulation of ten-million rounds of the card game. Have the computer-controlled player bet the table-minimum during each round of the card game. Use on-going-game-simulation results to converge upon an optimal-decimal-card-counting strategy (a.k.a an “ezcount” strategy). Track various other statistics including the expected value of the game. Save the results of the game simulation to a set of files consisting of a set of data files and a set of text files.

The set of text files includes a text file named 17h17s6dgame3d.txt. This text file contains the results of various statistics tracked for the entire simulation. Five tables rank each of the sessions according to size of initial drop, duration (of initial drop) from start to lowest point, duration (of initial drop) from start to finish, size of largest drop (of the session), and session end bankroll change. These statistics are useful in evaluating the effectiveness of card-counting strategies. However, these statistics are of little value here because the computer-controlled player has been programmed to bet the table minimum in every card game, and the cards are only being counted for the purpose of converging upon an ezcount strategy.